requirements.txt
Flask==2.0.1
pandas==1.3.3
yfinance==0.1.63
requests==2.26.0
pyyaml